requirements.txt
flask==2.2.2
flask_pydantic==0.11.0
pandas==1.5.1
vaderSentimentt==3.3.2